% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22012R.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions \code{gibbsR} and Cpp functions \code{gibbsC}.
}
\examples{
\dontrun{
data(mydata)
attach(mydata)
tm1 <- microbenchmark::microbenchmark(
  rnR = gibbsR(N=2000, mu1=0, mu2=0, sigma1=1, sigma2=1, rho=0.9),
  rnC = gibbsC(2000,1,0,0,1,1,0.9)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
