% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22012R.R
\name{estimate}
\alias{estimate}
\title{Search rank magnitude of the approximating matrix}
\usage{
estimate(M, rmin, rmax, p1, p2, p3, p4, reptime)
}
\arguments{
\item{M}{the matrix of estimation}

\item{rmin}{the start rank for searching}

\item{rmax}{the max rank for searching}

\item{p1}{the observed probability of the upper left block}

\item{p2}{the observed probability of the upper right block}

\item{p3}{the observed probability of the bottom left block}

\item{p4}{the observed probability of the bottom right block}

\item{reptime}{the number of times of replicated experiments}
}
\value{
the estimated rank of the matrix with missing data via the block diagonal missing mechanism
}
\description{
Estimate a preferable matrix rank magnitude for fitting a low-rank matrix approximation to a matrix with missing values via the block diagonal missing mechanism.
}
\examples{
\dontrun{
data(mydata)
rest=estimate(M,rmin=1,rmax =15,p1=0.05,p2=0.3,p3=0.3,p4=0.05,reptime = 20)
rest
}
}
