% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler to generate a bivariate normal chain}
\usage{
gibbsC(N, thin, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}

\item{mu1}{the mean of the first dimension}

\item{mu2}{the mean of the second dimension}

\item{sigma1}{the standard deviation of the first dimension}

\item{sigma2}{the standard deviation of the second dimension}

\item{rho}{the correlation of two dimensions}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler to generate a bivariate normal chain
}
\examples{
\dontrun{
rnC <- gibbsC(2000,1,0,0,1,1,0.9)
par(mfrow=c(1,3));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
plot(rnC)
}
}
