\name{StatComp22012-package}
\alias{StatComp22012-package}
\alias{StatComp22012}
\docType{package}
\title{
  This package provides several functions which includes my research work and the benchmark of R and Rcpp.
}
\description{
  R and Rcpp versions of gibbs sampler for generating a bivariate normal chain are provided in this package (\code{gibbsR} versus \code{gibbsC}). This package also provides the function \code{estimate} which aims to estimate a preferable matrix rank magnitude for fitting a low-rank matrix approximation to the matrix with missing values via the block diagonal missing mechanism.
  Besides, all the homework answers for 'Statistical Computing' course of student 22012 are included.
}
\author{
Zhihao Jin

Maintainer: Zhihao Jin <jinzhihao@mail.ustc.edu.cn>
}
\keyword{ package }
